kafka,migrations,configs

+++ pkg/generator/user.go
@@ -4,14 +4,14 @@ import "github.com/brianvoe/gofakeit"
 
 type UserGenerator struct {
 	Name   string
-	Age    uint8
+	Age    int32
 	Gender string
 }
 
 func NewUserGenerator() *UserGenerator {
 	return &UserGenerator{
-		Name: gofakeit.BeerName(),
-		Age:  uint8(gofakeit.Int8()),
+		Name:   gofakeit.BeerName(),
+		Age:    gofakeit.Int32(),
 		Gender: gofakeit.Gender(),
 	}
 }
